+++ CHANGELOG.markdown
@@ -1,5 +1,6 @@
 0.3.0.0
 ----
 * Removed parser stuff
-* Added `Default`, `Semigroup`, `Monoid` and `Applicative` instances for `Layout`
+* Added `Default`, `Semigroup`, `Monoid`, and `Applicative` instances for `Layout`
 * Added "extractors" (`System.Directory.Layout.Lens`)
+* Added `Apply` and `Bind` instances for `Layout`